refactor: move memory exporter to v1 package
Move the export types and builder logic into a dedicated v1 package
to prepare for creating a v2 export format.

Changes:
- Create internal/storage/memory/export/v1/ package
- Move Export, Time, Entity types to v1/types.go
- Move Build() function and helpers to v1/builder.go
- Simplify export.go to use v1 package for building exports
- Add BuildExport() public method for thread-safe access
- Update tests to use v1.Export type
